Refactored in order to share code between bulk validation and AI scoring
